Create function caching utility
Create a decorator which tries to load the results of a function from a
local file if it exists. If the file does not exist, then the decorator
calls the function and caches the results to disk.

This is often necessary when interfacing with web APIs and other slow
processes.