Add function to retrieve all articles in a journal
Add the ability to retrieve all of the articles associated with a
journal by traversing the Elsevier sitemap. The sitemap stores articles
according to their "PII", a proprietary Elsevier article identifier.

An example URL that links directly to an article is:
http://api.elsevier.com/content/article/pii/S2352187216300018

The PII of an article can be used to retrieve the XML version (with
Elsevier's DTD) or a 'simplified' document which contains a structured
metadata 'header' and the body of the text itself as one big UTF-8
string, without markup.