Create asynchronous web helper functions
Create webscraping helper functions which retrieve a set of pages
asynchronously. Using Requests to retrieve a webpage is slow because the
program blocks all other events from occurring while waiting for the
remote server. By using an asynchronous process to perform the
retrieval, we can simultaneously fetch a larger number of pages, leading
to a significant speed increase.