Adds unit testing framework to application
Adds a bunch of test classes and methods to the application.
For now, the existing tests are limited to the preprocess module.

Directory src/tests used for storing unit tests.

Tests can be run (discovered) with the standard library's unittest module:
$ python3 -m unittest

Included a sample eLife XML article as setUp data for most of the tests.