Adds a multiprocess disease NER annotation runner
Runs DNorm (Pubtator suite) on parsed paragraphs from articles.

Can be run the same way as preprocess.chem_ner and preprocess.gene_ner

Because DNorm is a Java application, this script has a Java dependency as well.
(tested on Java 8)

preprocess.disease_ner requires 10GB of RAM per Java process. The number of
Java processes to be launched (in the multiprocessing pool) is set
automatically based on available physical RAM and available CPU cores.

Additional information is available in the module docstring.