Adds a module for running Stanford CoreNLP
preprocess.prep_corenlp takes an input directory of articles in parform format,
a user-specified output directory, and path to a Stanford CoreNLP installation.

Files are converted to the appropriate plain text format, split into batches,
and submitted to a PBS queue for processing with CoreNLP.

For use on PBS clusters.