Adds a module for parsing MEDLINE XML data
Abstracts will be extracted alongside their PMID identifiers and saved to a
given output directory. If a PMC-ids.csv ID mapping file is given, PMIDs will
be mapped to DOIs and files will be saved by DOIs (first choice) or PMIDs
(second choice) if DOIs are unavailable for an article.

MEDLINE abstracts can be downloaded from the NLM with an appropriate license.