Moves subdirectory organizing function to util
Creates a new util function that reorganizes a directory with a large number
of files. Subdirectories are created and files are moved such that each
subdirectory has no more than max_files_per_subdir files in it.

This is normally not an issue, but may be necessary when dealing with extremely
large numbers of files in individual directories (> 10000).

Also adds a unit test for this new function.