Adds DeepDive article loader and app structure
Adds an article loader script written for the DeepDive v0.8 framework.

articles.tsv.sh is required for the current version of DD to load the more
useful script articles.tsv.py

The Python script reads the bioshovel_config.json file and reads in gzipped/tar
archives of articles or abstracts, parses them for text, replaces newline
characters (which seem to be incompatible with the psql parameters used here),
and prints them to STDOUT so that they can be imported into the Postgres
database.

* db.url contains database connection information.
* app.ddlog only contains the table schema (in DDlog) for the articles table